modify cauchy stress intergration and add elastic_dynamic_cauchy documents in Extrasource
