+++ SPHINXsys/src/shared/particle_dynamics/solid_dynamics/elastic_dynamics.cpp
@@ -80,24 +80,6 @@ namespace SPH
 			stress_PK1_B_[index_i] = F_[index_i] * elastic_solid_.StressPK2(F_[index_i], index_i) * B_[index_i];
 		}
 		//=================================================================================================//
-		CauchyIntegration1stHalf::
-			CauchyIntegration1stHalf(BaseInnerRelation &inner_relation)
-			: Integration1stHalf(inner_relation)
-		{
-			particles_->registerVariable(cauchy_stress_, "CauchyStress");
-		}
-		//=================================================================================================//
-		void CauchyIntegration1stHalf::initialization(size_t index_i, Real dt)
-		{
-			pos_[index_i] += vel_[index_i] * dt * 0.5;
-			F_[index_i] += dF_dt_[index_i] * dt * 0.5;
-			rho_[index_i] = rho0_ / F_[index_i].determinant();
-			Matd inverse_F_T = F_[index_i].inverse().transpose();
-			Matd almansi_strain = 0.5 * (Matd::Identity()  - (F_[index_i] * F_[index_i].transpose()).inverse());
-			//obtain the Cauchy stress 
-			cauchy_stress_[index_i] = elastic_solid_.StressCauchy(almansi_strain, F_[index_i], index_i)* inverse_F_T * B_[index_i];
-		}
-		//=================================================================================================//
 		KirchhoffParticleIntegration1stHalf::
 			KirchhoffParticleIntegration1stHalf(BaseInnerRelation &inner_relation)
 			: Integration1stHalf(inner_relation){};
