+++ tests/user_examples/xjt_2d_cauchy_oscillating_beam/regression_test_tool/BeamObserver_Position.xml
@@ -1,0 +1,103 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<Position>
+    <Snapshot_0 Position_0="~[0.198877,6.69983e-20]" />
+    <Snapshot_440 Position_0="~[0.196991,0.0279851]" />
+    <Snapshot_880 Position_0="~[0.190874,0.0532234]" />
+    <Snapshot_1316 Position_0="~[0.182499,0.0750712]" />
+    <Snapshot_1746 Position_0="~[0.173532,0.0919238]" />
+    <Snapshot_2176 Position_0="~[0.164827,0.104567]" />
+    <Snapshot_2600 Position_0="~[0.159129,0.111272]" />
+    <Snapshot_3020 Position_0="~[0.157897,0.113198]" />
+    <Snapshot_3440 Position_0="~[0.161181,0.109547]" />
+    <Snapshot_3865 Position_0="~[0.167126,0.101598]" />
+    <Snapshot_4295 Position_0="~[0.175392,0.0881136]" />
+    <Snapshot_4725 Position_0="~[0.184791,0.070036]" />
+    <Snapshot_5164 Position_0="~[0.192925,0.0463404]" />
+    <Snapshot_5604 Position_0="~[0.197922,0.0199278]" />
+    <Snapshot_6044 Position_0="~[0.199028,-0.00786905]" />
+    <Snapshot_6484 Position_0="~[0.195571,-0.0345494]" />
+    <Snapshot_6924 Position_0="~[0.188994,-0.0597185]" />
+    <Snapshot_7358 Position_0="~[0.179653,-0.0804465]" />
+    <Snapshot_7788 Position_0="~[0.170323,-0.0964664]" />
+    <Snapshot_8218 Position_0="~[0.163316,-0.106639]" />
+    <Snapshot_8640 Position_0="~[0.159038,-0.112238]" />
+    <Snapshot_9060 Position_0="~[0.158398,-0.112462]" />
+    <Snapshot_9480 Position_0="~[0.161747,-0.108349]" />
+    <Snapshot_9910 Position_0="~[0.169278,-0.0981717]" />
+    <Snapshot_10340 Position_0="~[0.178473,-0.0832678]" />
+    <Snapshot_10770 Position_0="~[0.18745,-0.0631526]" />
+    <Snapshot_11209 Position_0="~[0.194811,-0.0393955]" />
+    <Snapshot_11649 Position_0="~[0.198534,-0.012719]" />
+    <Snapshot_12089 Position_0="~[0.198385,0.0154381]" />
+    <Snapshot_12529 Position_0="~[0.194012,0.0425622]" />
+    <Snapshot_12967 Position_0="~[0.186122,0.0661632]" />
+    <Snapshot_13397 Position_0="~[0.177444,0.0851413]" />
+    <Snapshot_13827 Position_0="~[0.168852,0.0992799]" />
+    <Snapshot_14256 Position_0="~[0.16137,0.10893]" />
+    <Snapshot_14676 Position_0="~[0.157777,0.11291]" />
+    <Snapshot_15096 Position_0="~[0.158929,0.112008]" />
+    <Snapshot_15520 Position_0="~[0.164184,0.105637]" />
+    <Snapshot_15950 Position_0="~[0.171831,0.0944748]" />
+    <Snapshot_16380 Position_0="~[0.180844,0.0782958]" />
+    <Snapshot_16813 Position_0="~[0.189611,0.0575136]" />
+    <Snapshot_17253 Position_0="~[0.195999,0.0320725]" />
+    <Snapshot_17693 Position_0="~[0.199161,0.00430134]" />
+    <Snapshot_18133 Position_0="~[0.197464,-0.0233829]" />
+    <Snapshot_18573 Position_0="~[0.19226,-0.0489294]" />
+    <Snapshot_19013 Position_0="~[0.18409,-0.0719048]" />
+    <Snapshot_19443 Position_0="~[0.174449,-0.0898566]" />
+    <Snapshot_19873 Position_0="~[0.165759,-0.103101]" />
+    <Snapshot_20297 Position_0="~[0.160291,-0.11037]" />
+    <Snapshot_20717 Position_0="~[0.158233,-0.113036]" />
+    <Snapshot_21137 Position_0="~[0.160142,-0.110508]" />
+    <Snapshot_21561 Position_0="~[0.165645,-0.103408]" />
+    <Snapshot_21991 Position_0="~[0.174057,-0.0906456]" />
+    <Snapshot_22421 Position_0="~[0.183439,-0.0729157]" />
+    <Snapshot_22859 Position_0="~[0.191946,-0.0499804]" />
+    <Snapshot_23299 Position_0="~[0.197535,-0.0243968]" />
+    <Snapshot_23739 Position_0="~[0.198885,0.00308205]" />
+    <Snapshot_24179 Position_0="~[0.196616,0.0306086]" />
+    <Snapshot_24619 Position_0="~[0.189854,0.0562639]" />
+    <Snapshot_25055 Position_0="~[0.180974,0.0778138]" />
+    <Snapshot_25485 Position_0="~[0.172311,0.0938676]" />
+    <Snapshot_25915 Position_0="~[0.164455,0.105372]" />
+    <Snapshot_26337 Position_0="~[0.159122,0.111577]" />
+    <Snapshot_26757 Position_0="~[0.157845,0.11308]" />
+    <Snapshot_27177 Position_0="~[0.161247,0.109114]" />
+    <Snapshot_27604 Position_0="~[0.168072,0.100124]" />
+    <Snapshot_28034 Position_0="~[0.177008,0.0857099]" />
+    <Snapshot_28464 Position_0="~[0.186136,0.067039]" />
+    <Snapshot_28902 Position_0="~[0.193512,0.0437344]" />
+    <Snapshot_29342 Position_0="~[0.198211,0.017083]" />
+    <Snapshot_29782 Position_0="~[0.198817,-0.011447]" />
+    <Snapshot_30222 Position_0="~[0.194751,-0.0381832]" />
+    <Snapshot_30662 Position_0="~[0.187941,-0.0625061]" />
+    <Snapshot_31092 Position_0="~[0.179054,-0.0819777]" />
+    <Snapshot_31522 Position_0="~[0.169642,-0.0977921]" />
+    <Snapshot_31952 Position_0="~[0.162224,-0.107632]" />
+    <Snapshot_32373 Position_0="~[0.158276,-0.112776]" />
+    <Snapshot_32793 Position_0="~[0.158875,-0.11203]" />
+    <Snapshot_33213 Position_0="~[0.162978,-0.107251]" />
+    <Snapshot_33643 Position_0="~[0.170332,-0.0966133]" />
+    <Snapshot_34073 Position_0="~[0.17914,-0.0817158]" />
+    <Snapshot_34508 Position_0="~[0.188258,-0.0607148]" />
+    <Snapshot_34948 Position_0="~[0.195579,-0.0360181]" />
+    <Snapshot_35388 Position_0="~[0.198793,-0.00871489]" />
+    <Snapshot_35828 Position_0="~[0.19811,0.0185827]" />
+    <Snapshot_36268 Position_0="~[0.193452,0.0450579]" />
+    <Snapshot_36708 Position_0="~[0.185063,0.0687857]" />
+    <Snapshot_37138 Position_0="~[0.175917,0.087549]" />
+    <Snapshot_37568 Position_0="~[0.167497,0.100984]" />
+    <Snapshot_37995 Position_0="~[0.161088,0.109463]" />
+    <Snapshot_38415 Position_0="~[0.158224,0.112807]" />
+    <Snapshot_38835 Position_0="~[0.159387,0.111467]" />
+    <Snapshot_39259 Position_0="~[0.164518,0.104778]" />
+    <Snapshot_39689 Position_0="~[0.17274,0.0927948]" />
+    <Snapshot_40119 Position_0="~[0.182368,0.075531]" />
+    <Snapshot_40554 Position_0="~[0.190818,0.0540369]" />
+    <Snapshot_40994 Position_0="~[0.196532,0.0289462]" />
+    <Snapshot_41434 Position_0="~[0.199202,0.00123652]" />
+    <Snapshot_41874 Position_0="~[0.196978,-0.0267475]" />
+    <Snapshot_42314 Position_0="~[0.191059,-0.0526407]" />
+    <Snapshot_42748 Position_0="~[0.182917,-0.0742355]" />
+</Position>
